[ADD] Dashboard for task count with filter for period and assignee.
dashboard view seem to be removed from 16.

1. in order to add dashboard using spreadsheet a json file added to mimic the spreadsheet file record can be found in the dashboard.xml
2. JSON file has the declarations for sheets in the dashboard and the filter in the dashboard view.
3. In order to make the user_ids in the task model filterable and more workable along with team data a sql view is implemented in report directory that joins project, team and user with task tables for data in the assignee sheet.

currently dashboard sheets should 10 rows in each table but can easily be increased, also dashboard can be beautified with charts and pivots view but required a lot of effort with JSON file. Instead of that there is an easier option by installing oca_speardsheet_dashboard module, any user should be able to create and add sheets to dashboard.